requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-environ==0.8.1
idna==3.3
Pillow==9.1.1
psycopg2==2.9.3
python-dotenv==0.20.0
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
